# clients/apple/Packages/TorrentCoreKit/Sources/TorrentCoreAPI/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivityLogLevel": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "format": "int32",
        "type": "integer"
      },
...
      "RuntimeSettingsDto": {
        "additionalProperties": false,
        "properties": {
          "appliedEngineEncryptionMode": {
            "nullable": true,
            "type": "string"
          },
          "appliedEngineMaximumConnections": {
            "format": "int32",
            "type": "integer"
          },
          "appliedEngineMaximumDownloadRateBytesPerSecond": {
            "format": "int32",
            "type": "integer"
          },
          "appliedEngineMaximumHalfOpenConnections": {
            "format": "int32",
            "type": "integer"
          },
          "appliedEngineMaximumUploadRateBytesPerSecond": {
            "format": "int32",
            "type": "integer"
          },
          "coldDownloadAbandonAfterHours": {
            "format": "int32",
            "type": "integer"
          },
          "coldDownloadRecoveryIntervalMinutes": {
            "format": "int32",
            "type": "integer"
          },
          "coldDownloadRecoveryThresholdMinutes": {
            "format": "int32",
            "type": "integer"
          },
          "completedTorrentCleanupMinutes": {
            "format": "int32",
            "type": "integer"
          },
          "completedTorrentCleanupMode": {
            "nullable": true,
            "type": "string"
          },
          "completionCallbackApiBaseUrlOverride": {
            "nullable": true,
            "type": "string"
          },
          "completionCallbackApiKeyOverride": {
            "nullable": true,
            "type": "string"
          },
          "completionCallbackArguments": {
            "nullable": true,
            "type": "string"
          },
          "completionCallbackCommandPath": {
            "nullable": true,
            "type": "string"
          },
          "completionCallbackEnabled": {
            "type": "boolean"
          },
          "completionCallbackFinalizationTimeoutSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "completionCallbackTimeoutSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "completionCallbackWorkingDirectory": {
            "nullable": true,
            "type": "string"
          },
          "deleteLogsForCompletedTorrents": {
            "type": "boolean"
          },
          "engineConnectionFailureLogBurstLimit": {
            "format": "int32",
            "type": "integer"
          },
          "engineConnectionFailureLogWindowSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "engineEncryptionMode": {
            "nullable": true,
            "type": "string"
          },
          "engineMaximumConnections": {
            "format": "int32",
            "type": "integer"
          },
          "engineMaximumDownloadRateBytesPerSecond": {
            "format": "int32",
            "type": "integer"
          },
          "engineMaximumHalfOpenConnections": {
            "format": "int32",
            "type": "integer"
          },
          "engineMaximumUploadRateBytesPerSecond": {
            "format": "int32",
            "type": "integer"
          },
          "engineRuntime": {
            "nullable": true,
            "type": "string"
          },
          "engineSettingsRequireRestart": {
            "type": "boolean"
          },
          "maxActiveDownloads": {
            "format": "int32",
            "type": "integer"
          },
          "maxActiveMetadataResolutions": {
            "format": "int32",
            "type": "integer"
          },
          "metadataRefreshRestartDelaySeconds": {
            "format": "int32",
            "type": "integer"
          },
          "metadataRefreshStaleSeconds": {
            "format": "int32",
            "type": "integer"
          },
          "partialFileSuffix": {
            "nullable": true,
            "type": "string"
          },
          "partialFilesEnabled": {
            "type": "boolean"
          },
          "retrievedAtUtc": {
            "format": "date-time",
            "type": "string"
          },
          "seedingStopMinutes": {
            "format": "int32",
            "type": "integer"
          },
          "seedingStopMode": {
            "nullable": true,
            "type": "string"
          },
          "seedingStopRatio": {
            "format": "double",
            "type": "number"
          },
          "supportsLiveUpdates": {
            "type": "boolean"
          },
          "updatedAtUtc": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "usesPersistedOverrides": {
            "type": "boolean"
          }
        },
        "required": [
          "appliedEngineMaximumConnections",
          "appliedEngineMaximumDownloadRateBytesPerSecond",
          "appliedEngineMaximumHalfOpenConnections",
          "appliedEngineMaximumUploadRateBytesPerSecond",
          "coldDownloadRecoveryIntervalMinutes",
          "coldDownloadRecoveryThresholdMinutes",
          "completedTorrentCleanupMinutes",
          "completedTorrentCleanupMode",
          "completionCallbackEnabled",
          "completionCallbackFinalizationTimeoutSeconds",
          "completionCallbackTimeoutSeconds",
          "deleteLogsForCompletedTorrents",
          "engineConnectionFailureLogBurstLimit",
          "engineConnectionFailureLogWindowSeconds",
          "engineMaximumConnections",
          "engineMaximumDownloadRateBytesPerSecond",
          "engineMaximumHalfOpenConnections",
          "engineMaximumUploadRateBytesPerSecond",
          "engineRuntime",
          "engineSettingsRequireRestart",
          "maxActiveDownloads",
          "maxActiveMetadataResolutions",
          "metadataRefreshRestartDelaySeconds",
          "metadataRefreshStaleSeconds",
          "partialFilesEnabled",
          "partialFileSuffix",
          "retrievedAtUtc",
          "seedingStopMinutes",
          "seedingStopMode",
          "seedingStopRatio",
          "supportsLiveUpdates",
          "usesPersistedOverrides"
        ],
        "type": "object"
      },
...
      "ServiceProblemDetailsDto": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "nullable": true,
            "type": "string"
          },
          "detail": {
            "nullable": true,
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "nullable": true,
            "type": "object"
          },
          "instance": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "target": {
            "nullable": true,
            "type": "string"
          },
          "title": {
            "nullable": true,
            "type": "string"
          },
          "traceId": {
            "nullable": true,
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "ServiceRestartRequestResultDto": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "nullable": true,
            "type": "string"
          },
          "requestedAtUtc": {
            "format": "date-time",
            "type": "string"
          },
          "serviceLabel": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "message",
          "requestedAtUtc",
          "serviceLabel"
        ],
        "type": "object"
      },
...
      "TorrentHistoryDetailDto": {
        "additionalProperties": false,
        "properties": {
          "callbackCompletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "callbackLastError": {
            "nullable": true,
            "type": "string"
          },
          "callbackStartedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "categoryKey": {
            "nullable": true,
            "type": "string"
          },
          "completionCallbackFeedback": {
            "$ref": "#/components/schemas/CompletionCallbackFeedbackDto"
          },
          "completionCallbackLabel": {
            "nullable": true,
            "type": "string"
          },
          "dataDeleted": {
            "type": "boolean"
          },
          "downloadCompletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "downloadRootPath": {
            "nullable": true,
            "type": "string"
          },
          "downloadStartedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "finalPayloadPath": {
            "nullable": true,
            "type": "string"
          },
          "infoHash": {
            "nullable": true,
            "type": "string"
          },
          "invokeCompletionCallback": {
            "type": "boolean"
          },
          "lastActivityAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "lastUpdatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "latestCallbackStatus": {
            "nullable": true,
            "type": "string"
          },
          "latestConnectedPeerCount": {
            "format": "int32",
            "type": "integer"
          },
          "latestDownloadRateBytesPerSecond": {
            "format": "int64",
            "type": "integer"
          },
          "latestDownloadedBytes": {
            "format": "int64",
            "type": "integer"
          },
          "latestErrorMessage": {
            "nullable": true,
            "type": "string"
          },
          "latestProgressPercent": {
            "format": "double",
            "type": "number"
          },
          "latestTorrentState": {
            "nullable": true,
            "type": "string"
          },
          "latestTotalBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "latestTrackerCount": {
            "format": "int32",
            "type": "integer"
          },
          "latestUploadRateBytesPerSecond": {
            "format": "int64",
            "type": "integer"
          },
          "latestUploadedBytes": {
            "format": "int64",
            "type": "integer"
          },
          "latestWaitReason": {
            "nullable": true,
            "type": "string"
          },
          "magnetUri": {
            "nullable": true,
            "type": "string"
          },
          "metadataResolvedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "nullable": true,
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/TorrentHistoryOutcome"
          },
          "removalKind": {
            "$ref": "#/components/schemas/TorrentRemovalKind"
          },
          "removalReason": {
            "nullable": true,
            "type": "string"
          },
          "removedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "removedByCleanupPolicy": {
            "type": "boolean"
          },
          "seedingStartedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "serviceInstanceIdLastSeen": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "torrentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "dataDeleted",
          "invokeCompletionCallback",
          "lastUpdatedAt",
          "latestConnectedPeerCount",
          "latestDownloadedBytes",
          "latestDownloadRateBytesPerSecond",
          "latestProgressPercent",
          "latestTorrentState",
          "latestTrackerCount",
          "latestUploadedBytes",
          "latestUploadRateBytesPerSecond",
          "magnetUri",
          "name",
          "outcome",
          "removedByCleanupPolicy",
          "submittedAt",
          "torrentId"
        ],
        "type": "object"
      },
      "TorrentHistoryOutcome": {
        "enum": [
          "Active",
          "Removed",
          "Abandoned"
        ],
        "type": "string"
      },
      "TorrentHistorySummaryDto": {
        "additionalProperties": false,
        "properties": {
          "categoryKey": {
            "nullable": true,
            "type": "string"
          },
          "dataDeleted": {
            "type": "boolean"
          },
          "downloadCompletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "downloadRootPath": {
            "nullable": true,
            "type": "string"
          },
          "downloadStartedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "infoHash": {
            "nullable": true,
            "type": "string"
          },
          "lastActivityAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "lastUpdatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "latestCallbackStatus": {
            "nullable": true,
            "type": "string"
          },
          "latestConnectedPeerCount": {
            "format": "int32",
            "type": "integer"
          },
          "latestDownloadRateBytesPerSecond": {
            "format": "int64",
            "type": "integer"
          },
          "latestDownloadedBytes": {
            "format": "int64",
            "type": "integer"
          },
          "latestErrorMessage": {
            "nullable": true,
            "type": "string"
          },
          "latestProgressPercent": {
            "format": "double",
            "type": "number"
          },
          "latestTorrentState": {
            "nullable": true,
            "type": "string"
          },
          "latestTotalBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "latestTrackerCount": {
            "format": "int32",
            "type": "integer"
          },
          "latestUploadRateBytesPerSecond": {
            "format": "int64",
            "type": "integer"
          },
          "latestUploadedBytes": {
            "format": "int64",
            "type": "integer"
          },
          "latestWaitReason": {
            "nullable": true,
            "type": "string"
          },
          "metadataResolvedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "name": {
            "nullable": true,
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/TorrentHistoryOutcome"
          },
          "removalKind": {
            "$ref": "#/components/schemas/TorrentRemovalKind"
          },
          "removalReason": {
            "nullable": true,
            "type": "string"
          },
          "removedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "removedByCleanupPolicy": {
            "type": "boolean"
          },
          "seedingStartedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "torrentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "dataDeleted",
          "lastUpdatedAt",
          "latestConnectedPeerCount",
          "latestDownloadedBytes",
          "latestDownloadRateBytesPerSecond",
          "latestProgressPercent",
          "latestTorrentState",
          "latestTrackerCount",
          "latestUploadedBytes",
          "latestUploadRateBytesPerSecond",
          "name",
          "outcome",
          "removedByCleanupPolicy",
          "submittedAt",
          "torrentId"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Management API for the TorrentCore service host.",
    "title": "TorrentCore Service API",
    "version": "v1"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/api/history": {
      "get": {
        "operationId": "History_GetAll",
        "parameters": [
          {
            "in": "query",
            "name": "TorrentName",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "CategoryKey",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "State",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "Outcome",
            "schema": {
              "$ref": "#/components/schemas/TorrentHistoryOutcome"
            }
          },
          {
            "in": "query",
            "name": "Removed",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "FromDate",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ToDate",
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "Take",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TorrentHistorySummaryDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "History"
        ]
      }
    },
    "/api/history/by-torrent/{torrentId}": {
      "get": {
        "operationId": "History_GetByTorrentId",
        "parameters": [
          {
            "in": "path",
            "name": "torrentId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentHistoryDetailDto"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "tags": [
          "History"
        ]
      }
    },
...
    "/api/host/restart-service": {
      "post": {
        "operationId": "Host_RestartService",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceRestartRequestResultDto"
                }
              }
            },
            "description": "Accepted"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "Host"
        ]
      }
    },
    "/api/host/runtime-settings": {
      "get": {
        "operationId": "Host_GetRuntimeSettings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeSettingsDto"
                }
              }
            },
            "description": "OK"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "Host"
        ]
      },
      "put": {
        "operationId": "Host_UpdateRuntimeSettings",
        "requestBody": {
          "content": {
            "application/*\u002Bjson": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuntimeSettingsRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuntimeSettingsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuntimeSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeSettingsDto"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Bad Request"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "Host"
        ]
      }
    },
...
    "/api/logs": {
      "get": {
        "operationId": "Logs_GetRecent",
        "parameters": [
          {
            "in": "query",
            "name": "take",
            "schema": {
              "default": 100,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "level",
            "schema": {
              "$ref": "#/components/schemas/ActivityLogLevel"
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "eventType",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "torrentId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "serviceInstanceId",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fromUtc",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "toUtc",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActivityLogEntryDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "Logs"
        ]
      }
    },
    "/api/logs/delete-orphaned-torrent-logs": {
      "post": {
        "operationId": "Logs_DeleteOrphanedTorrentLogs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteOrphanedTorrentLogsResultDto"
                }
              }
            },
            "description": "OK"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceProblemDetailsDto"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "tags": [
          "Logs"
        ]
      }
    },
...
  },
...
}